+++ ReleaseNotes/release_notes.md
@@ -1,5 +1,9 @@
 ## New features
 
+ 1. allow to define `amin` parameter (cut-off for alpha, top avoid vanishing of alpha-parameter) for `Ostap::Math::Needham` via
+    constructor
+
+
 ## Backward  incompatible 
 
 ## Bug fixes